Fix highlighting of number of matches
The syntax for specifying the number of matches looks similar to pest
block syntax, e.g.,

    ANY{2}

vs

    rule = { ANY* }

With that the previous syntax definition failed parsing of blocks containing
rules with given number of matches.

This patch makes it possible for `pestBlock` to contain other `pestBlock` so we
can match constructs with a number of matches inside rule blocks. While this
doesn't capture the exact semantics it still gives a satisfying highlighting.